Changed directory structure.
If git log and similar commands ends in a dead-end on this commit, you need
to pass --follow to follow the file(s) over rename boundaries. Github will
unfortunetly not display this properly yet.

This most likely breaks lots of the projects (now in Build/) and might break
some scripts. It's all for the better in the long run though.

Data folder has not been restructered yet in this commit.